refactor: refactor slipp.controller package
