+++ slipp/src/main/java/slipp/controller/user/UserSessionUtils.java
@@ -1,4 +1,4 @@
-package slipp.controller;
+package slipp.controller.user;
 
 import slipp.domain.User;
